split tweedand users
