+++ db.js
@@ -64,10 +64,11 @@ pool.on('error', function (err) {
  */
 
 
-// const allPokemonModelsFunction = require('./models/pokemon');
+
+
 const allTweedrModelsFunction = require('./models/tweedr');
 
-// const pokemonModelsObject = allPokemonModelsFunction( pool );
+
 const tweedrModelsObject = allTweedrModelsFunction( pool );
