+++ routes.js
@@ -16,16 +16,18 @@ module.exports = (app, allModels) => {
   // const pokemonControllerCallbacks = require('./controllers/pokemon')(allModels);
   const tweedrControllerCallbacks = require('./controllers/tweedr')(allModels);
 
+  const usersControllerCallbacks = require('./controllers/users')(allModels);
+
   app.get('/tweedr', tweedrControllerCallbacks.index);
   app.get('/', tweedrControllerCallbacks.index);
 
   app.get('/tweet', tweedrControllerCallbacks.tweet);
   app.post('/tweet', tweedrControllerCallbacks.tweetPost);
 
-  app.get('/login', tweedrControllerCallbacks.login);
-  app.post('/login', tweedrControllerCallbacks.loginPost);
+  app.get('/login', usersControllerCallbacks.login);
+  app.post('/login', usersControllerCallbacks.loginPost);
 
-  app.get('/register', tweedrControllerCallbacks.register);
-  app.post('/register', tweedrControllerCallbacks.registerPost);
+  app.get('/register', usersControllerCallbacks.register);
+  app.post('/register', usersControllerCallbacks.registerPost);
 
 };
